Add login action that updates config with new token
